Moves the path transition code into its own header and makes the scan class call it as well.
